LayoutTests/fast/frames/javascript-url-form-target-new-window.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test.js"></script>
<script>
description("Tests that javascript URL execute in the right context when set as action on a form with a target popup.");
jsTestIsAsync = true;

function didRunJSURLInContext(_contextName)
{
    contextName = _contextName;
    // Since the form has target="testPopup", the javascript URL should execute in the context of the popup, not the top frame.
    shouldBeEqualToString("contextName", "testPopup");
    finishJSTest();
}

onload = () => {
    window.name = "topWindow";
    document.getElementById("submitButton").click();
};
</script>
<form id="testAnchor" target="testPopup" action="javascript:opener.didRunJSURLInContext(window.name);">
<input id="submitButton" type="submit" value="Click me"></input>
</form>
</body>
</html>
